+++ src/simplification/Simplifier.ts
@@ -7,6 +7,7 @@ import simplifyItems from './SimplifyItems';
 import simplifyExtend from './SimplifyExtend';
 import { SimplificationOptions } from '../models/SimplificationOptions';
 import simplifyAdditionalProperties from './SimplifyAdditionalProperties';
+import { isModelObject } from './Utils';
 
 export class Simplifier {
   static defaultOptions: SimplificationOptions = {
@@ -130,20 +131,4 @@ export class Simplifier {
 export function simplify(schema : Schema | boolean) : CommonModel[] {
   const simplifier = new Simplifier();
   return simplifier.simplify(schema);
-}
-
-/**
- * check if CommonModel is a separate model or a simple model.
- */
-export function isModelObject(model: CommonModel) : boolean {
-  // This check should be done instead, needs a refactor to allow it though:
-  // this.extend !== undefined || this.properties !== undefined
-  if (model.type !== undefined) {
-    // If all possible JSON types are defined, don't split it even if it does contain object.
-    if (Array.isArray(model.type) && model.type.length === 6) {
-      return false;
-    }
-    return model.type.includes('object');
-  }
-  return false;
 }
