+++ test/simplification/additionalProperties.spec.ts
@@ -1,29 +1,18 @@
 /* eslint-disable no-undef */
-import * as fs from 'fs';
-import * as path from 'path';
 import { CommonModel } from '../../src/models';
-import { Simplifier} from '../../src/simplification/Simplifier';
+import {Simplifier} from '../../src/simplification/Simplifier';
 import simplifyAdditionalProperties from '../../src/simplification/SimplifyAdditionalProperties';
-
-let mockModule = null;
 jest.mock('../../src/simplification/Simplifier', () => {
-  mockModule = {
+  return {
     Simplifier: jest.fn().mockImplementation(() => {
       return {
-        simplifyRecursive: jest.fn(),
-        simplify: jest.fn()
+        simplifyRecursive: jest.fn().mockImplementation(() => [new CommonModel()]),
+        simplify: jest.fn().mockImplementation(() => [new CommonModel()])
       };
-    }),
-    isModelObject: jest.fn()
+    })
   };
-  return mockModule;
 });
 describe('Simplification to additionalProperties', () => {
-  beforeEach(() => {
-    // Clear all instances and calls to constructor and all methods:
-    mockModule.Simplifier.mockClear();
-    mockModule.isModelObject.mockClear();
-  });
   test('should always return undefined if type is not object', () => {
     const schema = {type: 'string'};
     const simplifier = new Simplifier();
@@ -32,7 +21,7 @@ describe('Simplification to additionalProperties', () => {
     const simplifiedAdditionalProperties = simplifyAdditionalProperties(schema, simplifier, commonModel);
     expect(simplifiedAdditionalProperties.newModels).toBeUndefined();
     expect(simplifiedAdditionalProperties.additionalProperties).toBeUndefined();
-    expect(mockModule.Simplifier).toHaveBeenCalledTimes(0);
+    expect(simplifier.simplify).toHaveBeenCalledTimes(0);
   });
   test('should always return undefined if all types has been defined', () => {
     const schema = {type: ['string', 'object', 'null', 'number', 'array', 'boolean']};
@@ -42,7 +31,7 @@ describe('Simplification to additionalProperties', () => {
     const simplifiedAdditionalProperties = simplifyAdditionalProperties(schema, simplifier, commonModel);
     expect(simplifiedAdditionalProperties.newModels).toBeUndefined();
     expect(simplifiedAdditionalProperties.additionalProperties).toBeUndefined();
-    expect(mockModule.Simplifier).toHaveBeenCalledTimes(0);
+    expect(simplifier.simplify).toHaveBeenCalledTimes(0);
   });
   test('should always return true if type object and nothing is defined', () => {
     const schema = {type: 'object'};
@@ -51,18 +40,8 @@ describe('Simplification to additionalProperties', () => {
     commonModel.type = 'object';
     const simplifiedAdditionalProperties = simplifyAdditionalProperties(schema, simplifier, commonModel);
     expect(simplifiedAdditionalProperties.newModels).toBeUndefined();
-    expect(simplifiedAdditionalProperties.additionalProperties).toEqual({
-      originalSchema: true,
-      type: [
-        'object',
-        'string',
-        'number',
-        'array',
-        'boolean',
-        'null',
-      ],
-    });
-    expect(mockModule.Simplifier).toHaveBeenCalledTimes(0);
+    expect(simplifiedAdditionalProperties.additionalProperties).toEqual({});
+    expect(simplifier.simplifyRecursive).toHaveBeenCalledTimes(1);
   });
   test('should return true if additionalProperties is true', () => {
     const schema = { type: 'object', additionalProperties: true};
@@ -71,18 +50,8 @@ describe('Simplification to additionalProperties', () => {
     commonModel.type = 'object';
     const simplifiedAdditionalProperties = simplifyAdditionalProperties(schema, simplifier, commonModel);
     expect(simplifiedAdditionalProperties.newModels).toBeUndefined();
-    expect(simplifiedAdditionalProperties.additionalProperties).toEqual({
-      originalSchema: true,
-      type: [
-        'object',
-        'string',
-        'number',
-        'array',
-        'boolean',
-        'null',
-      ],
-    });
-    expect(mockModule.Simplifier).toHaveBeenCalledTimes(0);
+    expect(simplifiedAdditionalProperties.additionalProperties).toEqual({});
+    expect(simplifier.simplifyRecursive).toHaveBeenCalledTimes(1);
   });
   test('should return false if additionalProperties is false', () => {
     const schema = { type: 'object', additionalProperties: false};
@@ -92,18 +61,15 @@ describe('Simplification to additionalProperties', () => {
     const simplifiedAdditionalProperties = simplifyAdditionalProperties(schema, simplifier, commonModel);
     expect(simplifiedAdditionalProperties.newModels).toBeUndefined();
     expect(simplifiedAdditionalProperties.additionalProperties).toBeUndefined();
-    expect(mockModule.Simplifier).toHaveBeenCalledTimes(0);
+    expect(simplifier.simplifyRecursive).toHaveBeenCalledTimes(0);
   });
   test('should return simplified additionalProperties if schema', () => {
     const schema = { type: 'object', additionalProperties: {type: 'object', properties: {test: {type: 'string'}}}};
-    const filename = path.resolve(__dirname, './additionalProperties/expected/simple.json');
-    const expectedSchemaString = fs.readFileSync(filename, 'utf8');
-    const expectedAdditionalProperties = JSON.parse(expectedSchemaString);
     const simplifier = new Simplifier();
     const commonModel = new CommonModel();
     commonModel.type = 'object';
     const simplifiedAdditionalProperties = simplifyAdditionalProperties(schema, simplifier, commonModel);
-    expect(simplifiedAdditionalProperties).toEqual(expectedAdditionalProperties);
-    expect(mockModule.Simplifier).toHaveBeenCalledTimes(1);
+    expect(simplifiedAdditionalProperties.additionalProperties).toEqual({});
+    expect(simplifier.simplifyRecursive).toHaveBeenCalledTimes(1);
   });
 });
